move posts to separate namespace and update README.md
