+++ server/.env.sample
@@ -3,3 +3,7 @@ SECRET_KEY='&*(OHDIDUOSIHDSA)'
 
 YELP_API_KEY = ""
 CLIENT_ID = ""
+
+FLASK_PORT = 5000
+FLASK_DEBUG = 1
+FLASK_HOST = "127.0.0.1"
